Bring open label field into another component
